feat: add authorization validation

+++ app/Http/Controllers/ChirpController.php
@@ -3,11 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\Models\Chirp;
+use Illuminate\Contracts\Auth\Guard;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Redirect;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -45,10 +49,7 @@ class ChirpController extends Controller
             ],
         );
 
-        \App\Models\Chirp::create([
-            "message" => $validated["message"],
-            "user_id" => null,
-        ]);
+        $request->user()->chirps()->create($validated);
 
         // Redirect back to the feed
         return redirect("/")->with("success", "Chirp created!");
@@ -67,6 +68,8 @@ class ChirpController extends Controller
      */
     public function edit(Chirp $chirp)
     {
+        $this->authorize("update", $chirp);
+
         return view("chirps.edit", compact("chirp"));
     }
 
@@ -75,6 +78,8 @@ class ChirpController extends Controller
      */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize("update", $chirp);
+
         $validated = $request->validate([
             "message" => "required|string|max:255",
         ]);
@@ -89,6 +94,7 @@ class ChirpController extends Controller
      */
     public function destroy(Chirp $chirp)
     {
+        $this->authorize("delete", $chirp);
         $chirp->delete();
 
         return redirect("/")->with("success", "Chirp deleted!");
